refactor: emotion type 통일
