+++ app/report/index.tsx
@@ -15,8 +15,10 @@ export default function Report() {
 이대로만 하자고."
 			/>
 			<MonthlyEmotion
-				mainColor="#ff685b"
-				subColor="#f3e9da"
+				emotion={{
+					mainColor: "#ff685b",
+					subColor: "#f3e9da",
+				}}
 				description="저번 달에 비해 '행복한' 키워드에 가까운 날이 더 많습니다."
 				onPress={() => {
 					router.push("/report/monthly-emotion");
@@ -31,15 +33,15 @@ export default function Report() {
 						count: 15,
 					},
 					{
-						label: "초조",
+						label: "근심",
 						mainColor: "#063FB2",
-						subColor: "#FF685B",
+						subColor: "#8529D4",
 						count: 8,
 					},
 					{
-						label: "초조",
-						mainColor: "#063FB2",
-						subColor: "#FF685B",
+						label: "무기력",
+						mainColor: "#8529D4",
+						subColor: "#F3E9DA",
 						count: 5,
 					},
 					{
