+++ features/record/ui/EmotionComplete.tsx
@@ -1,5 +1,5 @@
-import { EmotionChip } from "@/features/record/model/types";
 import * as S from "@/features/record/style/EmotionComplete.styles";
+import { EmotionChip } from "@/shared/types";
 import { CharacterBubble } from "@/shared/ui/CharacterBubble";
 import { EmotionCard } from "@/shared/ui/EmotionCard";
 import { useRouter, type Href } from "expo-router";
@@ -58,7 +58,7 @@ export const EmotionComplete: React.FC<EmotionCompleteProps> = ({
 					<EmotionCard
 						mainColor={selectedEmotion.mainColor}
 						subColor={selectedEmotion.subColor}
-						title={selectedEmotion.label}
+						title={selectedEmotion.label ?? ""}
 					/>
 				</Animated.View>
 			)}
