+++ features/report/ui/TopEmotion.tsx
@@ -1,14 +1,8 @@
 import * as S from "@/features/report/style/TopEmotion.styles";
 import RightIcon from "@/shared/assets/icons/right.svg";
 import { theme } from "@/shared/config/theme/theme";
+import { EmotionChip } from "@/shared/types";
 import { TouchableOpacity } from "react-native";
-
-interface EmotionChip {
-	label: string;
-	mainColor: string;
-	subColor: string;
-	count: number;
-}
 
 interface TopEmotionProps {
 	emotionChip: EmotionChip[];
