+++ shared/types/emotion.ts
@@ -1,0 +1,6 @@
+export interface EmotionChip {
+	label?: string;
+	mainColor: string;
+	subColor: string;
+	count?: number;
+}
