+++ shared/types/index.ts
@@ -1,1 +1,2 @@
 export * from "@/shared/types/avatar";
+export * from "@/shared/types/emotion";
